Move dot out of vclock.rs into dot.rs; lots of little fixes (clippy, etc..)
