+++ src/dot.rs
@@ -1,0 +1,37 @@
+use serde::{Deserialize, Serialize};
+use std::hash::{Hash, Hasher};
+
+use crate::Actor;
+
+/// Dot is a version marker for a single actor
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct Dot<A> {
+    /// The actor identifier
+    pub actor: A,
+    /// The current version of this actor
+    pub counter: u64,
+}
+
+impl<A: Actor> Dot<A> {
+    /// Build a Dot from an actor and counter
+    pub fn new(actor: A, counter: u64) -> Self {
+        Self { actor, counter }
+    }
+}
+
+impl<A: Copy> Copy for Dot<A> {}
+
+impl<A: PartialEq> PartialEq for Dot<A> {
+    fn eq(&self, other: &Self) -> bool {
+        self.actor == other.actor && self.counter == other.counter
+    }
+}
+
+impl<A: Eq> Eq for Dot<A> {}
+
+impl<A: Actor + Hash> Hash for Dot<A> {
+    fn hash<H: Hasher>(&self, state: &mut H) {
+        self.actor.hash(state);
+        self.counter.hash(state);
+    }
+}
