+++ src/gcounter.rs
@@ -1,8 +1,7 @@
 use num_bigint::BigUint;
 use serde::{Deserialize, Serialize};
 
-use crate::traits::{Causal, CmRDT, CvRDT};
-use crate::vclock::{Actor, Dot, VClock};
+use crate::{Actor, Causal, CmRDT, CvRDT, Dot, VClock};
 
 /// `GCounter` is a grow-only witnessed counter.
 ///
